// src/lib/api/generated/model/uploadRequestFileResult.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Myelin Base RAG API
 * API documentation for Myelin Base RAG backend
 * OpenAPI spec version: 1.0.0
 */
import type { UploadRequestFileResultStatus } from "./uploadRequestFileResultStatus";
import type { UploadRequestResult } from "./uploadRequestResult";

export interface UploadRequestFileResult {
  status: UploadRequestFileResultStatus;
  fileName: string;
  data?: UploadRequestResult;
  errorCode?: string;
}